+++ kotest-framework/kotest-framework-engine/src/commonMain/kotlin/io/kotest/engine/test/extensions/TimeoutTestExecutionExtension.kt
@@ -4,11 +4,17 @@ import io.kotest.core.config.configuration
 import io.kotest.core.test.TestCase
 import io.kotest.core.test.TestContext
 import io.kotest.core.test.TestResult
+import io.kotest.core.test.TestType
 import io.kotest.engine.TestTimeoutException
+import io.kotest.engine.events.invokeAfterInvocation
+import io.kotest.engine.events.invokeBeforeInvocation
 import io.kotest.engine.test.TimeoutExecutionContext
 import io.kotest.mpp.log
+import io.kotest.mpp.replay
+import io.kotest.mpp.timeInMillis
 import kotlinx.coroutines.TimeoutCancellationException
 import kotlinx.coroutines.withTimeout
+import kotlin.math.min
 
 /**
  * The [TimeoutExecutionContext] is used to provide a way of executing functions on the underlying platform
@@ -16,6 +22,7 @@ import kotlinx.coroutines.withTimeout
  */
 class TimeoutTestExecutionExtension(
    private val ec: TimeoutExecutionContext,
+   private val start: Long,
 ) : TestExecutionExtension {
 
    private fun resolvedTimeout(testCase: TestCase): Long =
@@ -23,6 +30,12 @@ class TimeoutTestExecutionExtension(
          ?: testCase.spec.timeout
          ?: testCase.spec.timeout()
          ?: configuration.timeout
+
+   private fun resolvedInvocationTimeout(testCase: TestCase): Long =
+      testCase.config.invocationTimeout?.inWholeMilliseconds
+         ?: testCase.spec.invocationTimeout()
+         ?: testCase.spec.invocationTimeout
+         ?: configuration.invocationTimeout
 
    override suspend fun execute(
       test: suspend (TestCase, TestContext) -> TestResult
@@ -34,12 +47,40 @@ class TimeoutTestExecutionExtension(
       val timeout = resolvedTimeout(testCase)
       log { "TestCaseExecutor: Test [${testCase.displayName}] will execute with timeout $timeout" }
 
+      // note: the invocation timeout cannot be larger than the test case timeout
+      val invocationTimeout = min(resolvedTimeout(testCase), resolvedInvocationTimeout(testCase))
+      log { "TestCaseExecutor: Test [${testCase.displayName}] will execute with invocationTimeout $invocationTimeout" }
+
       try {
          withTimeout(timeout) {
-            ec.executeWithTimeoutInterruption(timeout) { test(testCase, context) }
+            ec.executeWithTimeoutInterruption(timeout) {
+               // depending on the test type, we execute with an invocation timeout
+               when (testCase.type) {
+                  TestType.Container -> test(testCase, context)
+                  TestType.Test -> {
+                     // not all platforms support executing with an interruption based timeout
+                     // because it uses background threads to interrupt
+                     replay(
+                        testCase.config.invocations,
+                        testCase.config.threads,
+                        { testCase.invokeBeforeInvocation(it) },
+                        { testCase.invokeAfterInvocation(it) }) {
+                        ec.executeWithTimeoutInterruption(invocationTimeout) {
+                           withTimeout(invocationTimeout) {
+                              test(testCase, context)
+                           }
+                        }
+                     }
+                     TestResult.success(timeInMillis() - start)
+                  }
+               }
+            }
          }
       } catch (e: TimeoutCancellationException) {
-         throw TestTimeoutException(timeout, testCase.displayName)
+         when (testCase.type) {
+            TestType.Container -> throw TestTimeoutException(timeout, testCase.displayName)
+            TestType.Test -> throw TestTimeoutException(min(timeout, invocationTimeout), testCase.displayName)
+         }
       }
    }
 }
