+++ kotest-tests/kotest-tests-timeout/src/test/kotlin/com/sksamuel/kotest/timeout/TestCaseTimeoutListenerTest.kt
@@ -11,7 +11,6 @@ import io.kotest.core.test.TestCase
 import io.kotest.core.test.TestCaseConfig
 import io.kotest.core.test.TestContext
 import io.kotest.core.test.TestResult
-import io.kotest.engine.test.toTestResult
 import io.kotest.matchers.shouldBe
 import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.GlobalScope
